manage/requirements.txt
django
django-admin
django-mptt
django-mptt-admin
django-bootstrap-v5